+++ src/transaction.hpp
@@ -19,17 +19,17 @@ using std::string;
 
 class Transaction {
 public:
-	Transaction(const string& senderAddr, const string& recipientAddr, float amount, float fee, const string& content);
+	Transaction(const string& senderAddr, const string& recipientAddr, double amount, double fee, const string& content);
 	Transaction(
-		const string& senderAddr, const string& recipientAddr, time_t creationTime, time_t signTime, float amount,
-		float fee, const string& content, Signature* signature, const string& hash
+		const string& senderAddr, const string& recipientAddr, time_t creationTime, time_t signTime, double amount,
+		double fee, const string& content, Signature* signature, const string& hash
 	);
 
 	Receipt* GetReceipt();
 
 	string m_SenderAddr, m_RecipientAddr;
 	time_t m_CreationTime, m_SignTime;
-	float m_Amount, m_Fee;
+	double m_Amount, m_Fee;
 	string m_Content;
 
 	// A sender and reciever signature used for verifying the transaction
